pyproject.toml
[project]
name = "visiondistill"
version = "0.1.0"
description = "Distill foundational vision models (SAM2/SAM3) into compact YOLO models via pseudo-labeling"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Yajat" }]
keywords = ["distillation", "segmentation", "sam", "yolo", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "ultralytics>=8.0",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "tqdm>=4.60",
    "accelerate>=0.25",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
visiondistill = "visiondistill.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
